L'exposition comparait un bigint a du texte, et n'avait jamais tourne
Premier passage de la 04a contre une vraie Postgres : « operator does not exist:
bigint = text » sur la requete de fenetre. Elle n'a donc JAMAIS pu s'executer
ailleurs que contre la doublure, et 264 tests passaient au-dessus.

`users.id` et `matches.user_id` sont des `bigint` ; le pilote `pg` rend un bigint en
CHAINE, et le renvoie tel quel en parametre. Postgres infere alors `$1` en `text` et
refuse la comparaison. Le parametre est desormais type explicitement en `$1::bigint`,
aux deux endroits ou il porte un identifiant de joueur : la sous-requete qui borne la
fenetre au joueur, et la jointure qui la referme.

C'est le meme piege que `ledgerSolde`, qui rendait la chaine de `sum()` : un bigint
qui traverse le pilote n'est pas un nombre, et le dossier le paie pour la seconde
fois. Les deux autres defauts trouves par le job `db` ne sont pas traites ici.

+++ api/db-pg.js
@@ -264,9 +264,9 @@ const EXPOSITION_FENETRE_SQL = `
                and cree_le <= $5
                and motif = any($4::text[])
                and (compte_debit = any($3::text[]) or compte_credit = any($3::text[]))
-               and ${L.REFERENCE_BILLET_SQL} in (select id::text from matches where user_id = $1)) e
+               and ${L.REFERENCE_BILLET_SQL} in (select id::text from matches where user_id = $1::bigint)) e
       join matches m on m.id::text = e.billet
-     where m.user_id = $1`;
+     where m.user_id = $1::bigint`;
 
 // LE FUSIBLE GLOBAL, ET IL NE FAIT AUCUNE JOINTURE. Ce n'est pas un oubli : c'est un INTERRUPTEUR,
 // pas un invariant. Le rattacher à `matches` en ferait un agrégat non borné sur DEUX tables, et il
